pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lotus-ai"
version = "0.1.0"
description = "lotus"
readme = "README.md"
authors = [
    {name = "Liana Patel", email = "lianapat@stanford.edu"},
    {name = "Siddharth Jha", email = "sidjha@berkeley.edu"}
]
license = { text = "MIT License" }
requires-python = ">=3.9, <3.13"
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
]

dependencies = [
    "backoff>=2.2.1,<3.0.0",
    "faiss-cpu>=1.8.0.post1,<2.0.0",
    "numpy>=1.26.4,<2.0.0",
    "openai>=1.35.13,<2.0.0",
    "pandas>=2.2.2,<3.0.0",
    "sentence-transformers>=3.0.1,<4.0.0",
    "tiktoken>=0.7.0,<1.0.0",
    "tqdm>=4.66.4,<5.0.0",
]

[project.optional-dependencies]
docs = [
    "sphinx",
    "sphinx_rtd_theme",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
# Select a minimal set of rules
select = [
    "F",
    "E",
    "I",
]

ignore = [
    "E501",  # Line too long
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
docstring-code-format = false
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"**/{docs}/*" = ["ALL"]
"**__init__.py" = ["ALL"]
